% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{create_EGG_update}
\alias{create_EGG_update}
\title{Generates an EGG update file}
\usage{
create_EGG_update()
}
\description{
This function prompts the user to provide information about what they
did this week for the EGG group in Pitt's Urology Department.
}
\keyword{update}
