% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg.updates-package.R
\docType{package}
\name{egg.updates-package}
\alias{egg.updates}
\alias{egg.updates-package}
\title{egg.updates: Create Updates for Pitt's EGG Group}
\description{
Prompts user for input and automatically renders a word document in the style of weekly update templates for the EGG group at Pitt Urology.
}
\author{
\strong{Maintainer}: Alex Watts \email{acwatts1995@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
